requirements.txt
django>=1.2.3
-e svn+http://django-compress.googlecode.com/svn/trunk/#egg=django_compress
-e git+git://github.com/django-extensions/django-extensions.git#egg=django-extensions
-e git+git://github.com/robhudson/django-debug-toolbar.git#egg=django-debug-toolbar
-e git+git://github.com/jslang/zombie.git#egg=zombie
gunicorn
django-sentry